requirements.txt
Django==1.8.4
django-mptt==0.5.5
django-extensions>=1.2.5
djangorestframework>=3.0
django-assets==0.8
django-yamlfield==0.4
django_openstack_auth>=1.1.3
python-ceilometerclient>=1.0.6
python-cinderclient>=1.0.6
python-glanceclient>=0.9.0
python-heatclient>=0.2.3
python-keystoneclient>=0.7.0
python-neutronclient>=2.3.4,<3
python-novaclient==2.17.0
pytz>=2010h
six>=1.6.0
requests!=2.8.0,>=2.5.2
lesscpy>=0.9j
lockfile>=0.8
gunicorn==18.0
setproctitle==1.1.8
celery>=3.1.13
django-celery>=3.1.10
raven>=4.2.3
python-statsd==1.6.4

python-memcached==1.54
psycopg2==2.6.1
